+++ src/detectors/index.ts
@@ -16,7 +16,7 @@ import { largeComponentDetector } from "./largeComponent.js";
 import { largeFunctionDetector } from "./largeFunction.js";
 import { namingDriftDetector } from "./namingDrift.js";
 import { propDrillingDetector } from "./propDrilling.js";
-import { pythonDeadAbstractionDetector, pythonDuplicateLogicDetector, pythonTodoCommentDetector } from "./python.js";
+import { pythonDeadAbstractionDetector, pythonDuplicateLogicDetector, pythonTodoCommentDetector } from "./python/index.js";
 import { rnHostForwardingDetector } from "./rnHostForwarding.js";
 import { routeHandlerSizeDetector } from "./routeHandlerSize.js";
 import { routeSprawlDetector } from "./routeSprawl.js";